The Shape class
This commit adds the Shape class which represents a bounded piece of the
plane, with a boundary consisting of arbitrary paths. A Shape object can
be constructed from existing 2D primitives (Rect, Circle, Parallelogram,
Ellipse), but the most general constructor takes an arbitrary PathVector
and a "fill rule". This general constructor requires the rather involved
`flatten` algorithm to uncross the paths and remove any redundant parts.
Due to its complexity, the flatten function is implemented in a separate
file `shape-flatten.cpp`.

Although the API of the Shape class is still in its infancy, one already
useful member function is `area()`, which can correctly compute the area
of a shape (fill region), accounting for holes, any pieces surrounded by
holes, etc. In the future, boolean operations will be implemented on the
`Shape` objects.

The new functionality is extensively documented, including a stand-alone
page explaining the `flatten` algorithm.  The documentation of the Shape
class populates the previously empty "Shape Operations" section which is
accessible from the home page of the html docs.

Tests for the Shape class are found in the `shape-test.cpp` file.