Repair the intersection of conic sections
The intersect() function for conic sections was hopelessly broken and
is reimplemented in this commit. The new implementation uses the method
of resultants to eliminate one of the variables from the system of
equations and then solves for the other variable.

Although the solutions are found analytically, rounding errors are
inevitable. Therefore, the solutions are polished by the application of
a small number (max 8) of iterations of the Newton-Raphson method.